v0.5: multi points

+++ core/loss.py
@@ -8,8 +8,9 @@ import numpy as np
 
 
 def sad_loss(data: MattingData, gt):
-    return np.sum(np.abs(data.alpha_matte[data.isu] - gt[data.isu]) / 255) / 1000
+    return np.sum(np.abs(data.alpha_matte[data.isu].astype(float) - gt[data.isu].astype(float)) / 255) / 1000
 
 
 def mse_loss(data: MattingData, gt):
-    return np.sum(np.square((data.alpha_matte[data.isu] - gt[data.isu]) / 255)) / data.u_size
+    return np.sum(
+        np.square((data.alpha_matte[data.isu].astype(float) - gt[data.isu].astype(float)) / 255)) / data.u_size
